backend/requirements.txt
# Core Agent Framework
langchain==0.3.14
langchain-core==0.3.83
langchain-community==0.3.14
langchain-groq==0.2.1

# OpenAI (optional - for vision fallback)
openai>=1.0.0

# RAG & Vector Store
chromadb==0.5.5
sentence-transformers==3.0.1
PyMuPDF==1.24.9

# Backend
fastapi==0.112.1
uvicorn==0.30.6
python-multipart==0.0.9

# Database (vehicle profiles - PostgreSQL / SQLite fallback)
sqlalchemy==2.0.32
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite>=0.19.0

# Web Scraping
beautifulsoup4==4.12.3
selenium==4.18.1
requests==2.32.3

# Image Processing & Vision
Pillow==10.3.0
opencv-python==4.9.0.80

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0

# Keras compatibility fix for sentence-transformers
tf-keras

# Utilities
pydantic==2.8.2
python-dotenv==1.0.1

# Authentication & Security
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0

# Agentic AI Dependencies
langchain-experimental==0.0.65
wikipedia==1.4.0
